requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
dj-database-url==2.1.0
Django==4.2.6
django-celery-beat==2.5.0
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-jazzmin==2.6.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==21.2.0
inflection==0.5.1
install==1.3.5
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kombu==5.3.5
mysqlclient==2.2.0
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
phonenumbers==8.13.33
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.31.0
rpds-py==0.13.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
